python/1password/requirements.txt
click 
pyyaml
